refactor: 인터뷰 도메인 및 매퍼 수정 #36
- Interview 클래스에 likes 속성 추가
- InterviewMapper에 likes 매핑 로직 추가
