+++ src/modules/quiz/domain/interview-group.type.ts
@@ -1,5 +1,4 @@
-import { FindInterviewResDto } from '../dto/interview.res.dto';
-
+import { FindInterviewResDto } from '../application/dto/interview.res.dto';
 export type InterviewGroup = {
   mainCategoryName: string;
   subCategoryName: string;
